requirements.txt
kuksa-client == 0.4.3
pyyaml
